backend/data-service/pyproject.toml
[tool.poetry]
name = "data-service"
version = "0.1.0"
description = "Data service for chatbot system with BigQuery integration"
authors = ["Renzo Chiarino <rz.chiarino@gmail.com>"]
readme = "README.md"
packages = [{ include = "data_service" }]

[tool.poetry.dependencies]
python = ">=3.09,<3.14"
fastapi = "^0.115.0"
google-cloud-bigquery = "^3.34.0"
google-cloud-storage = ">=1.32.0,<3.0.0"
sqlalchemy = "^2.0.41"
psycopg2-binary = "^2.9.10"
python-multipart = "^0.0.20"
python-dotenv = "^1.1.0"
pydantic-settings = "^2.9.1"
uvicorn = "^0.34.3"
vertexai = "^1.9.0"
google-cloud = "^0.34.0"
google-cloud-firestore = "^2.21.0"
langchain-google-firestore = "^0.5.0"
langchain-google-vertexai = "^2.0.26"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
